requirements.txt
fastapi
uvicorn
fasttext
psycopg2-binary
